remove window. approach for security
